Änderungen Examine Bed
