+++ adventure_state.json
@@ -1,1 +1,1 @@
-{"loc":2,"inventory":[],"rooms":[{"name":"Entrance","description":"You are in the entrance of the dungeon.","entities":[3],"actors":[],"exits":{"North":1}},{"name":"Corridor","description":"You are in a dark corridor.","entities":[],"actors":[],"exits":{"East":2,"South":0}},{"name":"Chamber","description":"There is a bed in the chamber. The pillows are soft.","entities":[],"actors":[1],"exits":{"West":1}},{"name":"Treasure Room","description":"You found the treasure room!","entities":[1],"actors":[],"exits":{"South":2}}],"entities":{"2":{"name":"A copper coin","description":"An old, dirty copper coin.","aliases":["coin","copper"]},"3":{"name":"A vending machine","description":"The vending machine has a small slid for the coins. The display is too dirty to reveal what it sells.","aliases":["machine","vending"]},"1":{"name":"A stack of gold","description":"It consists for roughly some hundred shiny gold coins.","aliases":["gold","stack","pile"]}},"actors":{"1":{"name":"Goblin","description":"A small green goblin leans against a door to the north.","aliases":["goblin"]}},"active_events":[0],"events":[{"condition_id":2,"message":"The bed is made of soft wood and has a comfortable mattress. Below the pillow you find a copper coin","command_stack":[{"AddItemToRoom":2}]}],"conditions":[{"Location":2},{"CommandIs":{"Examine":"bed"}},{"And":[0,1]}],"file_name":"adventure_state.json"}
+{"loc":0,"inventory":[],"rooms":[{"name":"Entrance","description":"You are in the entrance of the dungeon.","entities":[3],"actors":[],"exits":{"North":1}},{"name":"Corridor","description":"You are in a dark corridor.","entities":[],"actors":[],"exits":{"South":0,"East":2}},{"name":"Chamber","description":"There is a bed in the chamber. The pillows are soft.","entities":[],"actors":[1],"exits":{"West":1}},{"name":"Treasure Room","description":"You found the treasure room!","entities":[1],"actors":[],"exits":{"South":2}}],"entities":{"3":{"name":"A vending machine","description":"The vending machine has a small slid for the coins. The display is too dirty to reveal what it sells.","aliases":["machine","vending"]},"2":{"name":"A copper coin","description":"An old, dirty copper coin.","aliases":["copper","coin"]},"1":{"name":"A stack of gold","description":"It consists for roughly some hundred shiny gold coins.","aliases":["stack","gold","pile"]}},"actors":{"1":{"name":"Goblin","description":"A small green goblin leans against a door to the north.","aliases":["goblin"]}},"active_events":[0],"events":[{"condition_id":2,"message":"The bed is made of soft wood and has a comfortable mattress. Below the pillow you find a copper coin","command_stack":[{"AddItemToRoom":2},{"DeActivateEvent":0},{"ActivateEvent":1}]},{"condition_id":2,"message":"Now that you have taken the coin, you glance down at an empty bed","command_stack":[]}],"conditions":[{"Location":2},{"CommandIs":{"Examine":"bed"}},{"And":[0,1]}],"file_name":"adventure_state.json"}
