--- src/action.rs
@@ -1,14 +1,0 @@
-use crate::direction::Direction;
-
-pub enum Action {
-    Move(Direction),
-    Look,
-    Take(String),
-    Drop(String),
-    Use(String),
-    Save,
-    Load,
-    Quit,
-    Inventory,
-    None,
-}
